Showing transactions per wallet
Transacations are displayed using recycler view. data propert is a delgate which updates only when data changed.
Total amount is calculated based on transactions.
Fetching and calculations are offloaded to background thread (using doAsync) and results are displayed on ui thread.

When backgroud thread is working, progress bar is visible.