TransactionListCallbacks are optional to implement
Instead of 'hard binding' the listener on onAttach/onDetach lifecycle events, not it is up to the client of TransactionListFragment to decide when to attach/detach callbacks.
It helps with cases when user wants to see 'read-only' transaction list (like in Month Overview).

MonthOverviewActivity doesn't need to implement empty interface any more which makes it cleaner to read.